+++ src/tree.rs
@@ -155,7 +155,7 @@ impl Tree {
         loop {
             trace!("setting key {:?}", key.as_ref());
             if let Ok(res) =
-                self.insert_inner(key.as_ref(), Some(value.clone()), &mut guard)?
+                self.insert_inner(key.as_ref(), Some(value.clone()),false,  &mut guard)?
             {
                 return Ok(res);
             }
@@ -165,7 +165,8 @@ impl Tree {
     pub(crate) fn insert_inner(
         &self,
         key: &[u8],
-        mut value: Option<IVec>,
+        value: Option<IVec>,
+        is_transactional: bool,
         guard: &mut Guard,
     ) -> Result<Conflictable<Option<IVec>>> {
         let _measure = if value.is_some() {
@@ -177,7 +178,11 @@ impl Tree {
         let View { node_view, pid, .. } =
             self.view_for_key(key.as_ref(), guard)?;
 
-        let mut subscriber_reservation = self.subscribers.reserve(&key);
+        let mut subscriber_reservation = if is_transactional {
+            None
+        } else {
+            Some(self.subscribers.reserve(&key))
+        };
 
         let (encoded_key, last_value) = node_view.node_kv_pair(key.as_ref());
 
@@ -201,15 +206,12 @@ impl Tree {
 
         if link.is_ok() {
             // success
-            if let Some(res) = subscriber_reservation.take() {
-                let event = if let Some(value) = value.take() {
-                    subscriber::Event::Insert {
-                        key: key.as_ref().into(),
-                        value,
-                    }
-                } else {
-                    subscriber::Event::Remove { key: key.as_ref().into() }
-                };
+            if let Some(Some(res)) = subscriber_reservation.take() {
+                let event = subscriber::Event::single_update(
+                    self.clone(),
+                    key.as_ref().into(),
+                    value,
+                );
 
                 res.complete(&event);
             }
@@ -370,28 +372,49 @@ impl Tree {
     pub fn apply_batch(&self, batch: Batch) -> Result<()> {
         let _cc = concurrency_control::write();
         let mut guard = pin();
-        self.apply_batch_inner(batch, &mut guard)
+        self.apply_batch_inner(batch, None, &mut guard)
     }
 
     pub(crate) fn apply_batch_inner(
         &self,
         batch: Batch,
+        transaction_batch: Option<Event>,
         guard: &mut Guard,
     ) -> Result<()> {
-        let peg = self.context.pin_log(guard)?;
+        let peg = if transaction_batch.is_none() {
+            Some(self.context.pin_log(guard)?)
+        } else {
+            None
+        };
+
         trace!("applying batch {:?}", batch);
-        for (k, v_opt) in batch.writes {
+
+        let mut subscriber_reservation = self.subscribers.reserve_batch(&batch);
+
+        for (k, v_opt) in &batch.writes {
             loop {
-                if self.insert_inner(&k, v_opt.clone(), guard)?.is_ok() {
+                if self.insert_inner(k, v_opt.clone(), transaction_batch.is_some(), guard)?.is_ok() {
                     break;
                 }
             }
         }
 
-        // when the peg drops, it ensures all updates
-        // written to the log since its creation are
-        // recovered atomically
-        peg.seal_batch()
+        if let Some(res) = subscriber_reservation.take() {
+            if let Some(transaction_batch) = transaction_batch {
+                res.complete(&transaction_batch);
+            } else {
+                res.complete(&Event::single_batch(self.clone(), batch));
+            }
+        }
+
+        if let Some(peg) = peg {
+            // when the peg drops, it ensures all updates
+            // written to the log since its creation are
+            // recovered atomically
+            peg.seal_batch()
+        } else {
+            Ok(())
+        }
     }
 
     /// Retrieve a value from the `Tree` if it exists.
@@ -461,7 +484,7 @@ impl Tree {
         loop {
             trace!("removing key {:?}", key.as_ref());
 
-            if let Ok(res) = self.insert_inner(key.as_ref(), None, &mut guard)? {
+            if let Ok(res) = self.insert_inner(key.as_ref(), None, false, &mut guard)? {
                 return Ok(res);
             }
         }
@@ -570,14 +593,11 @@ impl Tree {
 
             if link.is_ok() {
                 if let Some(res) = subscriber_reservation.take() {
-                    let event = if let Some(new) = new {
-                        subscriber::Event::Insert {
-                            key: key.as_ref().into(),
-                            value: new,
-                        }
-                    } else {
-                        subscriber::Event::Remove { key: key.as_ref().into() }
-                    };
+                    let event = subscriber::Event::single_update(
+                        self.clone(),
+                        key.as_ref().into(),
+                        new
+                    );
 
                     res.complete(&event);
                 }
@@ -761,9 +781,16 @@ impl Tree {
     ///
     /// // `Subscription` implements `Iterator<Item=Event>`
     /// for event in subscriber.take(1) {
-    ///     match event {
-    ///         sled::Event::Insert{ key, value } => assert_eq!(key.as_ref(), &[0]),
-    ///         sled::Event::Remove {key } => {}
+    ///     // Events occur due to single key operations,
+    ///     // batches, or transactions. The tree is included
+    ///     // so that you may perform a new transaction or
+    ///     // operation in response to the event.
+    ///     for (tree, key, value_opt) in &event {
+    ///         if let Some(value) = value_opt {
+    ///             // key `key` was set to value `value`
+    ///         } else {
+    ///             // key `key` was removed
+    ///         }
     ///     }
     /// }
     ///
@@ -1064,14 +1091,11 @@ impl Tree {
 
             if link.is_ok() {
                 if let Some(res) = subscriber_reservation.take() {
-                    let event = if let Some(new) = &new {
-                        subscriber::Event::Insert {
-                            key: key.as_ref().into(),
-                            value: new.clone(),
-                        }
-                    } else {
-                        subscriber::Event::Remove { key: key.as_ref().into() }
-                    };
+                    let event = subscriber::Event::single_update(
+                        self.clone(),
+                        key.as_ref().into(),
+                        new.clone()
+                    );
 
                     res.complete(&event);
                 }